+++ docs/CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Add alpha cutoff support to the PBR shader ([#756])
 * Basic renderer setup helper function ([#771])
 * Shape mesh generators ([#777])
+* Add core support for Prefabs ([#716])
+* Add shape prefab support ([#786])
 
 ### Changed
 * UI systems will now never overwrite your local `UiTransform` values ([#774])
@@ -56,6 +58,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#774]: https://github.com/amethyst/amethyst/pull/774
 [#777]: https://github.com/amethyst/amethyst/pull/777
 [#776]: https://github.com/amethyst/amethyst/pull/776
+[#716]: https://github.com/amethyst/amethyst/pull/716
+[#786]: https://github.com/amethyst/amethyst/pull/786
 
 ## [0.7.0] - 2018-05
 ### Added